+++ client/ipa-client-automount
@@ -40,12 +40,12 @@ from six.moves.urllib.parse import urlsplit
 # pylint: enable=import-error
 
 from optparse import OptionParser  # pylint: disable=deprecated-module
+
+from ipaclient.install import ipachangeconf, ipadiscovery
 from ipalib import api, errors
 from ipalib.install import sysrestore
 from ipalib.install.kinit import kinit_keytab
 from ipapython import ipautil
-from ipaclient import ipadiscovery
-from ipaclient import ipachangeconf
 from ipapython.ipa_log_manager import root_logger, standard_logging_setup
 from ipapython.dn import DN
 from ipaplatform.constants import constants
