+++ install/tools/ipa-replica-manage
@@ -33,6 +33,7 @@ from six.moves.urllib.parse import urlparse
 from six.moves.xmlrpc_client import MAXINT
 # pylint: enable=import-error
 
+from ipaclient.install import ipadiscovery
 from ipapython import ipautil
 from ipaserver.install import replication, dsinstance, installutils
 from ipaserver.install import bindinstance, cainstance, certs
@@ -43,7 +44,6 @@ from ipalib.util import has_managed_topology, verify_host_resolvable
 from ipapython.ipa_log_manager import root_logger, standard_logging_setup
 from ipapython.dn import DN
 from ipapython.config import IPAOptionParser
-from ipaclient import ipadiscovery
 from ipaplatform.paths import paths
 
 # dict of command name and tuples of min/max num of args needed
