+++ ipaclient/install/ntpconf.py
